create isAudioFile() helper function
it returns whether the provided filename is of a supported audio format.

as mentioned in the code, lunar only currently supports .mp3 files although the audio library also supports other three formats.

in case the other formats get supported in lunar in the future, I've included an alternative code that should include the other file extensions.